api case section status
