+++ api/v2/cases/routes.js
@@ -1,6 +1,7 @@
 const inputValidations = require('./validations/input')
 module.exports = (server) =>{
   const handlers = require('./handlers')(server)
+  const getCasebyId = require('../../cases/route_prerequesites').getCasebyId(server)
   const CheckRoleCreate = require('../../users/route_prerequesites').CheckRoleCreate(server)
   const countCaseByDistrict = require('../../cases/route_prerequesites').countCaseByDistrict(server)
   const countCasePendingByDistrict = require('../../cases/route_prerequesites').countCasePendingByDistrict(server)
@@ -21,6 +22,19 @@ module.exports = (server) =>{
         ]
       },
       handler: handlers.CreateCase
+    },
+    {
+      method: 'GET',
+      path: '/v2/cases/{id}/status',
+      config: {
+          auth: 'jwt',
+          description: 'get specific case status',
+          tags: ['api', 'cases'],
+          pre: [
+              getCasebyId,
+          ]
+      },
+      handler: handlers.GetCaseSectionStatus
     },
   ]
 }
